requirements.txt
torch>=2.0
torchvision>=0.15
scikit-learn>=1.3
matplotlib>=3.7
tqdm>=4.65
Pillow>=9.5
PyYAML>=6.0
numpy>=1.24
